+++ docs/changelog.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Incorrect results when using `Deanonymize` multiple times ([#82](https://github.com/protectai/llm-guard/pull/82), thanks to [@andreaponti5](https://github.com/andreaponti5))
 
 ### Changed
-- `NoRefusal` scanner relies on the proprietary model [laiyer/distilroberta-base-rejection-v1](https://huggingface.co/laiyer/distilroberta-base-rejection-v1).
+- `NoRefusal` scanner relies on the proprietary model [ProtectAI/distilroberta-base-rejection-v1](https://huggingface.co/ProtectAI/distilroberta-base-rejection-v1).
 - `NoRefusal` support `match_type` parameter to choose between `sentence` and `all` matches.
 - Using `structlog` for better logging.
 - **[Breaking]**: `Code`: using new model [philomath-1209/programming-language-identification](https://huggingface.co/philomath-1209/programming-language-identification) with more languages support and better accuracy. Please update your `languages` parameter.
@@ -78,7 +78,7 @@ _0.3.5 and 0.3.6 were skipped due to build issues._
 
 ### Changed
 - Upgraded `json_repair` library ([issue](https://github.com/protectai/llm-guard/issues/44))
-- Use proprietary prompt injection detection model [laiyer/deberta-v3-base-prompt-injection](https://huggingface.co/laiyer/deberta-v3-base-prompt-injection)
+- Use proprietary prompt injection detection model [ProtectAI/deberta-v3-base-prompt-injection](https://huggingface.co/ProtectAI/deberta-v3-base-prompt-injection)
 
 ## [0.3.2] - 2023-11-15